Dialer: near by places and google dialer (2/2)
First of all we want to say big big big thanks to awesome developer
Xiao-Long ChenXiao-Long Chen who spend some time to get this working
with googles API. Just incredible result. So again huge thanks to him.

We work now as well quiet some time and tested a lot on it...and here
it is

Basically this patch brings into AOSP dialer an open source variant of
the famous google nearby places feature known from hammerhead stock builds. It relies on the
google services API. We thought short time about if we maybe go away from
googles API and use another one. But tbh no database/service API is that
powerfull then googles. Android is google so we will provide it as it is.

Based on the incrediable work of Xiao-Long we just fixed smaller things
adapted it to our integrations and made it TRDS compatible.

PS2: restore + rebase

Change-Id: I8f08e49943e413d11125cde191761b59faa75fe8
Signed-off-by: Dan Trifan <jimsth@gmail.com>
Conflicts:
	res/values/colors.xml
Signed-off-by: DHOMD <DHO@dho.im>